requirements.txt
alembic==1.4.2
astroid==2.4.1
bcrypt==3.1.7
blinker==1.4
cffi==1.14.0
click==7.1.2
cryptography==2.9.2
dnspython==1.16.0
email-validator==1.1.0
flake8==3.8.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-User==1.0.2.2
Flask-WTF==0.14.3
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
passlib==1.7.2
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pylint==2.5.2
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.16
toml==0.10.0
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
